WebApp_Volunteers/requirements.txt
flask
flask-sqlalchemy
sqlalchemy
python
arrow
python-dotenv
wtforms